partial paths for whitelist urls
So I am using 
https://formidablepro.com/knowledgebase/formidable-api/
and it had api urls in the form at the bottom.  We are linking in from another application with shared auth (CAS) and require the user be logged in to access some forms using your plugin for this requirment.  We extract the data using formidables api and this is where the issue lies.  We can add white list entries for each set or form apis but it will be a pain to manage.   This adds base path filters to allow api access to the form apis.


 * yoursite.com/wp-json/frm/forms GET Get a list of forms
 * yoursite.com/wp-json/frm/forms POST Create a form
 * yoursite.com/wp-json/frm/forms/5 GET Get form #5 (Change 5 to the ID of your form)
 * yoursite.com/wp-json/frm/forms/5 DELETE Permanently delete form #5 (Change 5 to the ID of your form)
 * yoursite.com/wp-json/frm/forms/5/html GET Get the form HTML to display on a page
 * yoursite.com/wp-json/frm/forms/5/fields GET Get an array of fields from form #5
 * yoursite.com/wp-json/frm/forms/5/entries GET Get the entries from form #5
 * yoursite.com/wp-json/frm/forms/5/entries DELETE Delete the entries from form #5
 * yoursite.com/wp-json/frm/entries POST Create an entry
 * yoursite.com/wp-json/frm/entries/25 GET Get entry #25 in an array
 * yoursite.com/wp-json/frm/entries/25 POST/PUT/PATCH Update entry #25
 * yoursite.com/wp-json/frm/entries/25 DELETE Delete entry #25